Add workaround script for LILYPOND_DATADIR bug.
When I came to re-test the main build script for this font, I found
that another thing that's bit-rotted in the last few years is that my
technique for overriding $LILYPOND_DATADIR doesn't work reliably any
more, because of what looks like a bug in Lilypond itself - it ignores
your input value of that env variable, because it overwrites it on
startup with a value inferred from its own binary path.

It's possible to work around this with a creative wrapper shell
script, so for the moment, let's do that.

(The wrapper script is complicated further by the need to dodge
Debian's own wrapper script. See the comment in it for details...)