Add workaround tool for SVG scientific-notation bug.
A user reported last week that Lilypond fails to extract some glyphs
(notably the sharp sign) from the Gonville SVGs, because some of the
path strings contain float literals in scientific notation (1.2e-3 and
so on), which Lilypond's SVG-font-parsing regex rejects.

I've reported the bug to the Lilypond maintainers, and it looks as if
it's likely to be fixed at some point. However, a lot of people will
still be using older versions of Lilypond for a while yet (including
me!), so for pragmatism, I'm now postprocessing the SVG font files
through a textual filter that fixes up those float literals after they
come out of Fontforge.