Move font definitions out into their own file.
Now glyphs.py is just the main program that deals with rendering the
font into various output formats, and the actual _font design_ lives
in font.py. This also means a lot of stray global variables used
during glyph setup won't be left in the global namespace for the
output code.