Parallelise all the slow gs/potrace runs.
Now we're in a nice modern version of Python, this is pretty easy -
all we have to do is 'import multiprocessing' and take advantage of
the handy facilities it already provides. I was afraid I'd have to do
this by breaking the Python script up into tiny pieces and building a
Makefile, but there's really no need!