Fix goof in allocation of Unicode points for braces.
Ahem. Of course, just _after_ I hit push, I found that the brace
between two staves in the demo image was absurdly tiny. Looks as if
that was a trivial bug I introduced at the very last minute with the
parallelisation changes: I broke a loop up into two, and in the second
half, reused the iteration variable that was only updated in the first
half.