Script to preview a glyph from a Lilypond-format font.
This outputs Postscript that draws the glyph itself, but also marks
the origin, and the formal bounding box and attachment point from the
font's LILC entry. It can be applied to the standard Lilypond font, or
the Lilypond output of this font, which is useful for checking that
I've put attachment points in sensible places.