Update test instructions.
I tried to use the 'convert' command from the comment in test_glyph
recently and found that it doesn't work any more: ImageMagick has
changed what '-recolor' means, in a way I wasn't easily able to
compensate for. Replaced the command with a completely different one
that also has the effect I wanted.