font.py: refactor to use Python decorators.
Every glyph in this font is implemented by defining a Python function
which returns a GlyphContext, and then calling that function. These
days, we don't have to do that by writing a function call after each
function: we can use a decorator, which has the advantage that you can
write multiple decorators and pass it extra parameters.