Doubled G-clef.
A glyph added in newer versions of Lilypond: two G-clefs overlapping
each other apparently means the same thing as a G-clef with an 8 below
it, i.e. an octave lower than usual.

I've used a clipping path to remove some parts of the left clef, as if
it were hiding behind the right one. I think it looks less cluttered
that way.