Separate slashed from reversed turn.
I always thought that you notated an inverted turn (e.g. c-b-c-d-c) by
drawing a vertical stroke through the turn symbol. So that's what I
implemented as 'scripts.reverseturn' in Gonville. But current Lilypond
uses that symbol name for a _mirrored_ version of the glyph, and the
one with a vertical line is 'scripts.slashturn'. Now we match that.