Add more subdivisions of a quaver.
These don't seem to be in the new stable Lilypond 2.20 release, but
Lilypond 2.21 has introduced three further subdivisions of a quaver,
with 6, 7 and 8 tails, hence representing 2^-8, 2^-9 and 2^-10 of a
semibreve.

I'm constrained in my design of the tail glyphs because Lilypond is
dictating the length of the stem for each one - it doesn't seem to be
a parameter that the font file can control. Fortunately, it turns out
that the lengths work reasonably if I just use the exact same tail
style that I already had for 5-tailed quavers, and replicate it more
times.

The corresponding quaver rests are slanted a bit less away from
vertical, so they can squash together more conveniently. In Lilypond's
standard font that variation of slant is a gradual theme throughout
all the sub-quaver rests, but I think it looks nicer to keep them all
parallel until it really becomes impractical, so I've left the
existing five as they were.