Rewrite README for the modern world.
Now I put the Lilypond 2.20 font switching system front and centre,
and relegate the old horrors to 'previously, in 2.18'.

I've also removed discussion of LILYPOND_DATADIR, because I've found
that it's actually broken in 2.18. There are workarounds, but they're
quite horrible to start with, and more so when Debian wraps the actual
Lilypond binary with a script:

https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=951801

Instead, I've included a suggestion about unshare and bind mounts (the
same trick I use in Buildscr here) to try to help people. But really,
I'm going to have to start advising people to use the up-to-date LP if
they want to use Gonville without undue pain. It's a shame that 2.18
is still in current distros, but that'll pass.