Rework the output font directories.
Now that a stable Lilypond release (2.20) is out with a proper font-
switching mechanism, users no longer need to overwrite the previous
font in their Lilypond data dir, or switch the whole directory out for
a new one.

This commit changes the behaviour of glyphs.py --lily so that it
generates two different output directories, for different installation
methods.

'lilyfonts' is for new Lilypond: it's designed to be overlaid on top
of the Lilypond data dir, so that the font files in its 'otf' and
'svg' subdirs drop into the corresponding subdirs of
/usr/share/lilypond/2.20/fonts (or wherever). Then you can typeset a
score in Gonville by selecting it using the set-global-fonts system.

'lilyfonts-old' is for old Lilypond. That one is designed to _replace_
the fonts subdirectory of the Lilypond data dir, and it provides all
the Gonville font files under the same names that Lilypond 2.18 was
going to look for its standard font.