Replace lilysetup with a Lilypond 2.20 'stylesheet'.
For Lilypond 2.20, I'm going to start recommending that people
actually use the new font-switching system. That's configured in the
Lilypond input file, so you can make your own scores come out in
Gonville by writing an appropriate font definition statement in your
.ly files.

But what about separating style from content? If I want to pass a
score file on to other people, I might not want to force them all to
install Gonville. Conversely, if I _get_ a score from other people,
I'd rather be able to typeset it in my own choice of font without
having to edit the input file (especially if I'm keeping it up to date
via source control).

So it's nicer if you can have a command-line option to switch fonts.
Fortunately, you can: make an include file containing the appropriate
Lilypond setting, and then it can be pre-included on the command line
via 'lilypond -d include-settings=gonville.ily'.

Even easier is if I've already provided the include file myself. So in
this commit I add a suitable 'gonville.ily' and arrange to put it in
the distribution archive.

(You could think of that file as a sort of 'stylesheet'; in fact, the
one I use for my own actual score-writing includes a lot of other
personal preferences in addition to changing the notation font.)

I'm also removing the 'lilysetup' script, because there's no need to
be tricky about it any more. With the new layout of the lilyfonts-old
directory, all you have to do is to 'cp -rT' to replace files in an
original Lilypond data directory with the new ones.

Buildscr now depends on Lilypond 2.20 itself, because it builds the
demo images using the new switching stylesheet. In order to avoid
having to overwrite a real LP data directory, I've used the 'unshare'
mechanism to bind-mount a modified one over the top. I'll mention that
in a forthcoming README update.