Buildscr: replace -DNOBRACE with -DFASTBRACE.
This will stop me from making the same mistake again that I just fixed
in the previous commit: now we _do_ build the brace font in the faster
dev build cycle, we just don't put 576 different sizes of brace in it
any more. So if I do anything that makes the brace-font build actually
fail, I'll notice in early development.