Fix off-centre attachment on tenor G clef.
I thought this looked a little off when I first drew it, but forgot to
go back and take a closer look. Turns out I'd just got my maths wrong.
The stave spacing is 5 points, i.e. 5/72 inch, and in the normal
coordinate scale of this font, that's 5/27 * 1900. So two staves come
to twice that, which is 263.8888... pixels. But I'd written 268, which
I suspect was just a transcription error.