Fix small version of tenorised G clef.
All the other 'small' versions of clefs are simply the original clef
shrunk. (They're used when you change clefs in the middle of a line.)

But the tenorised G clef is an exception, because its spiral is
supposed to be centred on one stave line (representing a G), and its
upper protrusions are supposed to bracket another stave line
(representing the D four notes higher). So if you just shrink it in
the obvious way, then the distance between those two vertical
positions becomes wrong by a factor of 0.8.

Instead, the small tenorised G clef must be built in the same way as
the large one, but starting with scaled-down versions of the
components. Then the vertical distance is preserved.

(Incidentally, I claimed in the log message of commit 521eb894cb that
the protrusions borrowed from the variant C clef were supposed to
bracket middle C. That made so much logical sense to me that I didn't
actually check it. But in fact it's nonsense, because two stave lines
above a G is not a C, it's a D! In this clef - as you'd expect from a
treble clef transposed down by an octave - middle C is in a gap
between stave lines, not on a stave line at all.)