Shorten the variant segno by one loop.
When I came to test it in a real score, I immediately noticed that it
looked a bit too tall. But it took me longer to spot the reason why,
even though it should have been obvious: the components of the glyph
were fine, but I had just repeated the looping motif in the middle one
time too many compared to existing examples.

Reduced by one loop, it's now the same height as Lilypond's version.