Make the dots bigger in the fermata family.
A user suggested that the larger dots in Lilypond's default font are
actually a usability improvement, because in particular they make the
fermata dot and the staccato dot clearly distinct (even before you
look at the arc over the top of the former).

While I was here, I also aligned the bottom edge of the dot in each
fermata sign with the bottom edges of the stroke ends. Previously they
were off by about 3 design units, and I don't think that was on
purpose.

I also made fermata3 a bit higher, to avoid the enlarged dot hitting
its head on the ceiling.