New helper function, Curve.compute_nib_bounds.
This allows me to query a point on the curve and determine not just
the logical position there, but the bounding box of the pen nib when
it reaches that point of the curve. Useful for computing the bounds of
the actual ink at points on the curve, which I'm going to need for
font metrics in the upcoming SMuFL output mode.