Give the SMuFL fonts a version number.
Now that --ver actually works again, it makes a difference that I left
it out of the 'glyphs.py --smufl' invocation, whereas I put it in for
the Lilypond runs. Reinstated.

Also, I've put a version number in the zip file name too. Not doing
that in the previous commit was a mistake, but fortunately, I've only
given the URL to a few people for test purposes so far, and warned
them that it won't be permanent.

I've also stopped providing the individual SMuFL font files as
separate build products, because I'd have to rename them to have
version numbers in, and I don't think that's particularly useful.
Easier to have a zip file containing files with standard names (so
that when you unpack it, you don't have to adjust whatever was
pointing at the file), but version the zip file itself.