Make glyphs.py --ver actually work!
Because the primary option spelling is --ver, it sets a field called
'ver' in the output object of argparse. But I was retrieving
args.verstring rather than args.ver, so that whatever you passed to
--ver was ignored.

That would normally have caused a Python exception to let me know of
the mistake, except that I also set a default value for the wrong
field 'verstring', so that it did have a value to retrieve. Memo to
self: using parser.set_defaults() risks this kind of spelling mismatch
in a way that adding default='foo' to the add_argument call does not.

As a result, none of my font files actually had a version number. Now
fixed, ahem.